Create Classroom.py
This is a simple problem that allows a teacher to keep track of her math, reading, art, and computer activities. The teacher can add more students if needed and extend the parameter of the number of chair remaining at the activity tables.  As of now, there are 16 students with a max allowance of 4 students to a table.